fn: rename from fnctl
