+++ examples/hello/go/README.md
@@ -4,15 +4,15 @@ This example will show you how to test and deploy Go (Golang) code to IronFuncti
 
 ```sh
 # create your func.yaml file
-fnctl init <YOUR_DOCKERHUB_USERNAME>/hello
+fn init <YOUR_DOCKERHUB_USERNAME>/hello
 # build the function
-fnctl build
+fn build
 # test it
-cat hello.payload.json | fnctl run
+cat hello.payload.json | fn run
 # push it to Docker Hub
-fnctl push
+fn push
 # Create a route to this function on IronFunctions
-fnctl routes create myapp /hello
+fn routes create myapp /hello
 ```
 
 Now you can call your function on IronFunctions:
@@ -23,5 +23,5 @@ curl -H "Content-Type: application/json" -X POST -d @hello.payload.json http://l
 
 ## Dependencies
 
-Be sure you're dependencies are in the `vendor/` directory and that's it. 
+Be sure you're dependencies are in the `vendor/` directory and that's it.
