+++ fn/.gitignore
@@ -1,0 +1,3 @@
+fn
+vendor/
+/fn.exe
