+++ fn/entrypoint.sh
@@ -4,4 +4,4 @@ HOST=$(/sbin/ip route|awk '/default/ { print $3 }')
 
 echo "$HOST default localhost localhost.local" > /etc/hosts
 
-/fnctl "$@"
+/fn "$@"
