+++ fn/errors.go
