+++ fn/lambda.go
@@ -37,7 +37,7 @@ func lambda() cli.Command {
 	return cli.Command{
 		Name:      "lambda",
 		Usage:     "create and publish lambda functions",
-		ArgsUsage: "fnctl lambda",
+		ArgsUsage: "fn lambda",
 		Subcommands: []cli.Command{
 			{
 				Name:      "create-function",
