+++ fn/main.go
@@ -11,11 +11,11 @@ import (
 
 func main() {
 	app := cli.NewApp()
-	app.Name = "fnctl"
+	app.Name = "fn"
 	app.Version = "0.1.0"
 	app.Authors = []cli.Author{{Name: "iron.io"}}
 	app.Usage = "IronFunctions command line tools"
-	app.UsageText = `Check the manual at https://github.com/iron-io/functions/blob/master/fnctl/README.md
+	app.UsageText = `Check the manual at https://github.com/iron-io/functions/blob/master/fn/README.md
 
 ENVIRONMENT VARIABLES:
    API_URL - IronFunctions remote API address`
