--- fnctl/.gitignore
@@ -1,3 +1,0 @@
-fnctl
-vendor/
-/fnctl.exe
